Define the main class
